wip: Hacky attempt at fcgl
This works a bit better than what was already there but A) it's very
invasive, and B) it doesn't do the right thing if both -Fc and -Fo are
passed in. We probably have to actually wire up the DXIL backend to
have a mode where it *doesn't emit DXIL* to do this right.