[Transforms][DXIL] Introduce DXIL ShaderModel utility
This adds an abstraction for bidirectional translation between
DXIL-metadata and the modern representation of the DXIL shader model.
We use the utility in the DXILUpgrade pass and the DirectX backend's
lowering of the DXIL metadata.

In the DirectX backend this is mostly NFC, except it replaces some
asserts with errors. These should mostly be caught earlier in a normal
compiler flow, but if we do get here it's better to error than assert.