[DirectX] Remove new-pm versions of DXILResource passes. NFC
This is a bit of a hack, but these aren't actually used at the moment
and they'll cause naming conflicts with DXIL resource passes that I'm
adding to llvm/Analysis for replacing all of this with the target
extension type based solution. I'd rather leave the existing stuff in
place until we're ready to replace it completely, and this is the
simplest way to get away with that.

Pull Request: https://github.com/llvm/llvm-project/pull/100698